requirements.txt
# MagicFinance — Python Dependencies
# =====================================
# Install with: pip install -r requirements.txt
# Recommended: Python 3.11+

# ─── Core ──────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ─── Reddit scraping ───────────────────────────────────────────────────────────
praw>=7.7.0

# ─── Vector database (Nanobot VPS Qdrant) ─────────────────────────────────────
qdrant-client>=1.9.0

# ─── LLM inference (Ollama local API) ─────────────────────────────────────────
requests>=2.31.0

# ─── Market data & portfolio math ──────────────────────────────────────────────
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0
scipy>=1.12.0

# ─── Notebooks & visualisation ─────────────────────────────────────────────────
jupyter>=1.0.0
ipykernel>=6.0.0
matplotlib>=3.8.0
